change converter name
